% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_database.R
\name{database_init}
\alias{database_init}
\title{Get MLB data from 2015 to the present}
\usage{
database_init(sqlite_db_file)
}
\arguments{
\item{sqlite_db_file}{a place to write the SQLite database}
}
\description{
Scrapes the GameDay database and saves output to an SQLite database
}

