% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_database.R
\name{gameday_database_update}
\alias{gameday_database_update}
\title{Update GameDay database}
\usage{
gameday_database_update(sqlite_db_file)
}
\arguments{
\item{sqlite_db_file}{the SQLite database file to update}
}
\description{
updates the GameDay database
}

