% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_database.R
\name{get_stattleship_game_logs}
\alias{get_stattleship_game_logs}
\title{Get Stattleship game logs}
\usage{
get_stattleship_game_logs()
}
\value{
a Stattleship list of lists with the game logs
}
\description{
Fetch all (many pages) Stattleship MLB game logs
}

