% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_database.R
\name{get_stattleship_team_game_logs}
\alias{get_stattleship_team_game_logs}
\title{Get Stattleship team game logs}
\usage{
get_stattleship_team_game_logs()
}
\value{
a Stattleship list of lists with the team game logs
}
\description{
Fetch all (many pages) Stattleship MLB team game logs
}

