% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_database.R
\name{install_github_packages}
\alias{install_github_packages}
\title{Install GitHubPackages}
\usage{
install_github_packages()
}
\description{
installs packages from GitHub for baseball. Current packages are
\itemize{
\item stattleship/stattleship-r
\item rmscriven/retrosheet
\item cboettig/Sxslt
\item beanumber/openWAR}
}

