% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draftkings_utilities.R
\name{read_template}
\alias{read_template}
\title{Read DraftKings lineup template}
\usage{
read_template(file, contest_date)
}
\arguments{
\item{file}{a DraftKings lineup template file}

\item{contest_date}{the date of the contest (character, 'yyyy-mm-dd')}
}
\value{
a data frame with the player data
}
\description{
DraftKings provides a lineup template for each slate of games.
This function extracts the player data from that template.
}

