% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_database.R
\name{set_stattleship_token}
\alias{set_stattleship_token}
\title{Set Stattleship token}
\usage{
set_stattleship_token(token_file)
}
\arguments{
\item{token_file}{a file with the token}
}
\description{
The Stattleship API requires a token to work.
"set_stattleship_token" reads this token from a user-specified file.
See http://developers.stattleship.com.
}

